refactor(pipeline): extract digest formatter helpers
